server/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
python-multipart>=0.0.9
ebooklib>=0.18
beautifulsoup4>=4.12
azure-storage-blob>=12.19
openai>=1.30
pydantic>=2.0
